Added protection to destroy so that if it is called twice it does nothing This removes the protection from the implementation on every redefinition of destroy from EV_ANY_IMP
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@43901 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/vision2/interface/kernel/ev_any.e
@@ -93,7 +93,9 @@ feature -- Command
 			-- Destroy underlying native toolkit object.
 			-- Render `Current' unusable.
 		do
-			implementation.destroy
+			if not is_destroyed then
+				implementation.destroy
+			end
 		ensure
 			is_destroyed: is_destroyed
 		end
